feat(llm-cli): centralize cli binary resolution logic
- add shared binary resolver utilities for env, PATH, and fallback path resolution
- refactor Codex adapter to use the shared resolver while keeping compatibility behavior
- update llm_cli AGENTS guidance with optional env usage and provider implementation checklist

Made-with: Cursor
